Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> (1 comment)
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > spring framework provides configurable, easy to apply caching
> > > https://docs.spring.io/spring-framework/docs/current/reference/html/integration.html#cache
> > > 
> > > current implementation contains hard-coded configuration and no cache invalidation mechanism.
> > > not reliable
> > 
> > not sure what do you mean by not "reliable".. 
> > But indeed we could Caffeine, which has better performance, and supports annotations. that I guess is what you want
> 
> my bad didn't see the put for cache update.
> 
> anyway there are still the issues with this implementation:
> 
> 1) the cache serves the get schema context by dataspace and schema set name
> which is a rare case, which returns data for ui (exact case is still unknown yet);
> from other hand for frequent expected case - get schema context by dataspace and anchor name -
> this cache still requires extra db query, to find out which schema set is needed..
> it means we need to cache anchor to schema set name mapping as well, then only this 
> cache became useful, then this should be implemented as a separate artifact, not mixed with
> service logic
> 
> 2) the cache behavior is quite inflexible, it's unclear how to adopt it for different scenarios
> taking into account the schema set and/or data update frequency, 
> . the caching should be configurable

Hi Ruslan on the point of what we are caching, we cannot cache by Anchor as an anchor can be re-created with different schemas set. So we do rely on a call to the persistence layer to get the current schema set name/id. It will be a frequent call I agree but it is small and will rely on the built in optimization of the DB for this. But the schema Ste is 'immutable' and it makes sense to cache this (actually much longer then the initial proposal, but w can tune that later)

Patch-set: 3
Reviewer: Gerrit User 6639 <6639@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>