Update patch set 7
Patch Set 7:

(1 comment)

Not sure if I missed the grooming for this. In general, i would like to discuss the idea of having a cache as a separate and optional component, like we have the DB choice open. A vendor may like to plugin an existing cache(Redis etc) or have his own implementation for it. The DB choice could in some ways influence this too. We could def have a reference impl, but seems we should discuss this more.

Patch-set: 7
Reviewer: Gerrit User 6862 <6862@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>