Update patch set 9
Patch Set 9:

Niamh, this is just a quick README. All documentation for developers / user will go on doc.
If you consider that this README needs all those improvements is a must, please create a task for that. And someone can handle it as an independent task.
imho once we have a proper doc, this README should go away, so I wouldnt invest much time on it.
Regards

Patch-set: 9