Update patch set 10
Patch Set 10:

(4 comments)

I would agree with Niamh that it would be good to have a useful documentation in sync with the code.

It seems that we start having more comments in the review than in the readme file itself ;)
So, I'm proposing this patch #10 for the documenation.

Also, I'm stabilizing postgres version to 12.4 that is the ONAP recommended one for Honolulu:
https://wiki.onap.org/display/DW/Database%2C+Java%2C+Python%2C+Docker%2C+Kubernetes%2C+and+Image+Versions

Patch-set: 10
Reviewer: Gerrit User 5658 <5658@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>