Update patch set 9
Patch Set 9: Code-Review-1

(9 comments)

can we try to argue less and follow up constructive comment in a more positive manner please?

In generaal when a reviewer asks me for a better name, to extract something for additional clarification I just doing t as obviously that other developer felt the need for it even it was obvious to me, it is about seeing the code from the perspective of that other developer. Maybe especially because that reviewer is not as experienced or not as involved in the problem you are handling as you at this moment in time!

Patch-set: 9
Reviewer: Gerrit User 6639 <6639@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1