Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> while it seems ok as a standalone solution
> there will be problems for entire configuration co-existing with
> the one serving the db schema initialization.
> 
> due to there could be the only one liquibase configuration file
> per datasource it assumes the master configuration includes 
> schema configuration first the data preset then. so it assumes the 
> configuration for both should be allocated in same module - cps-ri.
> 
> i think there will be an issue if liquibase configuration parts 
> are delivered within different jars, so i'd move this configuration
> to a cps-ri module with appropriate folders renaming to clearly separate
> what belongs to schema and what's for data.
> 
> also i expected single sql file as a data carrier,
> splitting to multiple changesets seems extra for entire task.
> 
> also while the yang resources and data file a still being updated
> i'd reconsider liquibase approach in a favor of java code as more flexible 
> way to update the resources if required.

As per my understanding the data preload is required only while using cps-xnf-proxy. If some tenants decide not to use cps-xnf-proxy, this would not be required and hence it was agreed to keep the pre-load of data separate from the schema creation.
As far as the changesets are considered, keeping them separate helps in maintaining each of them and also that is suggested by liquibase to include a different changeset for each table.

Patch-set: 2