Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Duplicate for https://gerrit.onap.org/r/c/cps/+/118132
> 
> Please also see that I responded to comments made in 118132.

is not a duplicate, this commit fixes the root issue. Which is the capability to generate an image with wrong content.
Meanwhile your commit is about service name and build process(skip image generation), plus there still some issues presents. e.g. if you execute "mvn clean install -Djib.skip=false" using your commit, it will try to generate an image and will fail because missconfiguration.
When indeed, it should not even execute the plugin at all if not profile is selected.

And that is what this commit is addressing, your changes proposal can still go on top of this one, since they address different things.

Patch-set: 2