Update patch set 2
Patch Set 2:

We just need to  keep in mind that always generating the docker image also impacts `cps-maven-stage-master` job. By default, this job is also going to generate the docker image even if it is not is responsibility. `maven-docker-stage-master` job is responsible for that.

Then, `cps-maven-stage-master` will need to be changed if we want to remove docker image from it. And the same would apply to any process that does not need docker or is docker independent.

I'm not saying that we should not go that way if this is a CSIT requirement, this is just to highlight the impact.

Patch-set: 2
Reviewer: Gerrit User 5658 <5658@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>