Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > I'm wondering if this one could not also be a false positive if there is no security concern with the checksum we are generating for yang models.
> > > 
> > > I would imagine that a more robust hashing function could also require more resources and could take more time to compute (I don't have detailed numbers for this).
> > > 
> > > The tread-off I can see here is between the collision risk and performance. Performance being a key aspect for CPS.
> > > 
> > > What do you thing ?
> > 
> > Since this security hotspot is a very new addition to sonar, it could wait. Also agree with performance issues, though some numbers would be good.
> 
> In my opinion as we don't have proven performance bottleneck right now I think code for security is more important and I would prefer to chance the code as Aditya is proposing in this commit.

Yes, we have to put things on the balance to make good decisions. Unless I'm missing something, I don't think that we are solving a security issue here. With this change, it seems to me that we are reducing the risk of collision for 2 yang models to have the exact same checksum. And this checksum is then used for model management and not in any security context.

Sonar asked ourselves to answer these questions in order to take our decision:

---
Ask Yourself Whether

The hashed value is used in a security context like:

* User-password storage.
* Security token generation (used to confirm e-mail when registering on a website, reset password, etc ...).
* To compute some message integrity.

There is a risk if you answered yes to any of those questions.
---

It seems to me that we could respond no to all of them. Am I correct ?

Patch-set: 2
Reviewer: Gerrit User 5658 <5658@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>