Update patch set 1
Patch Set 1: Code-Review+1

I'm good with this change, but not be ready to be merged as it is not independant. It is on top on timestamps for docker image tag, for which there is a pending question.

Do ypou want to make it independant or wait ?

Patch-set: 1
Reviewer: Gerrit User 5658 <5658@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1