Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Based on Bruno's comment, maybe we should wait until after last fixes are merged or RC0 or RC1?!

As this is in master branch, if there is an issue we’ll face it for Istambul. This change should not go in honolulu branch.

Maybe what we can do is to wait after our latest honolulu release and make a test release in master to validate the change. This way we’ll know that we’ll be ready when it is time for releasing Istambul.

Patch-set: 1
Reviewer: Gerrit User 5658 <5658@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>