Update patch set 1
Patch Set 1:

Here is an extract of a discussion with Rishi about which registry to retrieve the image from either onap nexus or docker.io:

We also have a pipeline to build internally. It uses Bell internal registry by specifying the maven environment variable . It is currently:

   -Dbase.image=<bell-registry>/openjdk:11.0.9-jre-slim

But it would be cleaner if we could add a new variable `base-image-registry` or something like that. Its default value would be empty. In pom.xml, we could define:

   <base.image>${base-image-registry}onap/integration-java11:8.0.0</base.image>

It would give the option avoid docker.io and use onap nexus if needed, and for our Bell internal pipeline it would ensure that we are still aligned with cps default base image and we are only changing the registry.

Patch-set: 1
Reviewer: Gerrit User 5658 <5658@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>