Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > (1 comment)
> > 
> > > Patch Set 1:
> > > will you do NCMP REST in a separate commit?
> > 
> > yes, i'd like to provide an update after the module rename is merged
> 
> update: no ncmp update is expected for this task
> due to anchor is administrative entity. the ncmp manages data only.

NCMP needs o be able to add and remove xNF functions each of them is represented by an anchor (or CmHandle in NCMP) So both creation and deletion of CmHandle need to be supported and replace the hard-coded model and anchor that is now in use. 
However as the CmHandle/Anchor in NCMP is currently hardcoded and we need to discuss with E2E slicing team how to migrate away from the hard-coded solution it should NOT be part of the scope of this user story.  I'll update the scope now

Patch-set: 3
Reviewer: Gerrit User 6639 <6639@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>