Update patch set 1
Patch Set 1:

(2 comments)

This change is good to me. Only the typo with the where condition needs to be changed.

You can also test it locally with these steps:

---
1. Run a brain new postgres instance container on your local machine

  docker run --name postgres -p 5432:5432 -d \
    -e POSTGRES_DB=cpsdb -e POSTGRES_USER=cps -e POSTGRES_PASSWORD=cps \
    postgres:13.2-alpine

2. Checkout honolulu branch and cd to "cps-ri"

3. Run "status" liquibase script to observe that your database is not up to date and see the list of all changes set that need to be applied for honolulu

  mvn compile org.liquibase:liquibase-maven-plugin:4.3.1:status \
    -Dliquibase.url=jdbc:postgresql://localhost:5432/cpsdb \
    -Dliquibase.username=cps \
    -Dliquibase.password=cps \
    -Dliquibase.changeLogFile=changelog/changelog-master.yaml

4. Run "update" liquibase script to apply all pending changes set and bring your database to honolulu version

  mvn compile org.liquibase:liquibase-maven-plugin:4.3.1:update \
    -Dliquibase.url=jdbc:postgresql://localhost:5432/cpsdb \
    -Dliquibase.username=cps \
    -Dliquibase.password=cps \
    -Dliquibase.changeLogFile=changelog/changelog-master.yaml

5. Run "status" liquibase script to validate that your db is now up to date for honolulu
(Run step 3 command)

6. Checkout master branch (or the branch having your liquibase fix) and cd to "cps-ri"

7. Run "status" liquibase script to see the list of all change sets that need to be applied for istambul (if any)
(Run step 3 command)
- If the script is failing because of a checksum validation, it indicates that a specific change set already applied has been modified resulting in the script being now inconsistent with the database.
- If the script completes successfully, it lists pending change sets to be applied (if any).

8. Run "update" liquibase script to apply all pending changes set and bring your database to istambul version
(Run step 4 command)

9. Run "status" liquibase script to validate that your db is now up to date for istambul
(Run step 3 command)
---

For more information about some other Maven liquibase goals: https://docs.liquibase.com/tools-integrations/maven/commands/home.html

We also need to keep in mind that for reverting, here, we are changing a previous change set again. It means that we are going to break change sets already in place in our dev databases for master branch. This is still the right thing to do as it is easy to reset our dev databases. This is not something that we want to face or support when operational databases with business data are going to migrate from honolulu to istambul.

Patch-set: 1
Reviewer: Gerrit User 5658 <5658@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Work-in-progress: false