Implement DMI Registration
Implementing registration of CM-Handles to NCMP from DMI

Added Groovy Tests for these changes.

CM Handles are Written to fragment tables

Moved NetworkCmProxyDataServiceImplSpec.groovy as it was not in the same
package as the class it was testing. Meaning it didn't cover this when
it came to code coverage

Deprecating old API's along with old API Methods

Issue-ID: CPS-442

Change-Id: I819b9bf65280b1d968d3b75ca5ef2f9eb5617579
Signed-off-by: DylanB95EST <dylan.byrne@est.tech>