Update patch set 6
Patch Set 6:

(1 comment)

I see the module_name and revision getting stored in yang_resources when a single schema set file is uploaded. Eg : cps-cavsta-onap-internal2021-01-28.yang
But when a zip file is saved with multiple files, it doesnt save the module_name and revision in the yang-resources table.

As this story is related to fetching/listing the modules, the code changes for storing the module_name and revision could be taken as a different task.

Patch-set: 6
Reviewer: Gerrit User 6738 <6738@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>