Update patch set 9
Patch Set 9:

(5 comments)

> Patch Set 9:
> 
> forgot to say on above comment, depending on other comment and time remaining they might be done as postcode freeze improvement ?!

Current logic indicates that any child nodes will be updated until the exception is thrown. If the exception is thrown on the first node to be updated, none of the other nodes will be updated. If it is thrown on the last node to be updated, all nodes before this will be updated. I could attempt to add this before the code freeze on Thursday, and make the process transactional, to roll back all changes if any exceptions are thrown. But, I would also like to commit the remove functionality and apply code updates based on reviews of that too.

Patch-set: 9