Update patch set 6
Patch Set 6:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > (7 comments)
> > 
> > Could you explain why you need a separate 'root operation' ?! Its not clear from this commit. Maybe its part of the 2nd part of this user story but I would like to have some idea now...
> 
> I can see a bit of differentiation in the original stud re 'root' update and non-root udpates but I final to see the exact need. ie. what changed your mind?!

I changed the implementation again. I was not happy with the first implementation because 
- the operation was tightly coupled with the notification feature
- it was confusing too. For example, the create action was sending UPDATE operation to notification Service. 

The Second Implementation, which Toine reviewed, was a little better because it was less confusing. But, it added lots of if/else to identify if it is root or non-root. Operation enum was also a little confusing as CHILD_NODE_CREATA becomes 'update' inside the enum itself.

In the latest implementation ( Third approach :) ), the operation and the action method in the dataServiceImpl class are in sync. Also, the logic to identify 'root' or 'non-root' has been moved at one place in NotificationService.

Let me know if you have other ideas to make the implementation better.

Patch-set: 6