Update patch set 10
Patch Set 10:

> Patch Set 10: Code-Review-1
> 
> > Patch Set 7: Code-Review-1
> > 
> > Handling of DmiRequestException and ServerNcmpException is missing in DMNetworkCmProxyRestExceptionHandler.
> > This is where we also should return 400 for client request errors.
> 
> I'm putting back -1 for 2 reasons:
> 
> 1. It seems that previous -1 comment above has not be addressed yet.
> 
> 2. The parent of this change is not from master. It is an old commit from https://gerrit.onap.org/r/c/cps/+/126608/. With this dependency, this change should be merged after 126608. It will also be important to rebase from master and ensure that no obsolete code from the other change is kept.

Myself and Toine discussed this and as both exceptions extend NcmpExcption DMNetworkCmProxyRestExceptionHandler can remain as is... if this is not the case we can arrange a call and discuss.
For point 1) I will have a look back through the comments. for point 2) I am going to create new reviews that are not related so this will be avoided.

Patch-set: 10