Update patch set 3
Patch Set 3:

Please find below some notes about the discussion we had with Joe.

The requirement for the http rest api is to return 400 (client error) for errors the client is responsible for and 500 (system server error) for errors operation support team is responsible for.

This is important for the client to receive correct and accurate feedback about errors that are happening. We don't want the client to open tickets to the support because an error he is responsible for has been reported as server error.

On the other side, operation support team who is monitoring the application and server errors does not want to be alerted and start investigation on errors caused by incorrect client input.

That being said, assuming that `DmiRequestException` is raised by an invalid json payload provided by the client, it is important that this exception is handled on the rest layer in `DMNetworkCmProxyRestExceptionHandler` to return 400 http code.

We can have another call on Monday if there are other different point of views about this.

Patch-set: 3
Reviewer: Gerrit User 5658 <5658@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>