Update patch set 7
Patch Set 7:

The original error from csit test is identified in https://jenkins.onap.org/job/cps-review-verification-maven-istanbul/35/console. It is caused by:

  14:07:31 Pulling ncmp-dmi-plugin (nexus3.onap.org:10003/onap/ncmp-dmi-plugin:latest)...
  14:07:31 manifest for nexus3.onap.org:10003/onap/ncmp-dmi-plugin:latest not found: manifest unknown: manifest unknown

is the docker snapshot repository (https://wiki.onap.org/display/DW/Required+Tools#RequiredTools-Server-BasedToolsandRepositories). It is not really a surprise that no latest dmi image is found in this repo. Even if an image is found, it would be an un-predictable one. It is not suitable for integration testing.

Hence, the suggestion to use proxy repo and dmi plugin release ... To be reviewed and discussed more ...

Patch-set: 7