Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (7 comments)
> 
> > Patch Set 1:
> > 
> > (4 comments)
> > 
> > Also remove the Jira ID from the main commit message as it is already include in the Issue ID.
> > CSIT failing for register CM handle
> 
> CSIT tests passing on my local

Hi Anjella, I think I have found the issue. SDNC does not start for me as it exits due to it not finding certifivcatres. I assume you have these locally and that is why it works for you but not for me or on Jenkins. See below:

2022-02-04 17:11:07,870 - root - INFO - Connected status: True
2022-02-04 17:11:07,872 - root - ERROR - Error while extracting zip file(s). Exiting Certificate Installation.
2022-02-04 17:11:07,872 - root - INFO - Error details : [Errno 13] Permission denied: '/opt/opendaylight/current/certs/keys0'
Stoppping SDNR container due to failure in installing Certificates

Patch-set: 3
CC: Gerrit User 7743 <7743@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>