Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review-1
> 
> (1 comment)
> 
> Noticed an issue that all cm handles are returned if you use the following request body similar to this one, as if the request body is empty.
> 
> {
>   "publicCmHandleProperties": [
>     {
>       "Book": "Harry Potter",
>       "Author": "JK Rowling"    }
>   ]
> }

Might get teams opinion on this as I am struggling to implement validation for it, the json parser reads this as an empty list and as such returns all of the cmhandles, I am inclined to say it is an edge case that we can ensure will be avoided with good documentation. Otherwise I will have to completely change the structure of the openapi input object to validate the "cmHandleQueryParameters" is in the request body. This issue also currently exists in cps, it is not added by my code. Opinions from team are appreciated :)

Patch-set: 6