Update patch set 7
Patch Set 7:

Encountered a bug in the cmhandles id-searches endpoint that if no body is supplied it returns an anchor instead of all the cmhandles. Checked master as well and the bug already exists before this commit. Since this commit is so big already I will make a bug and it can be addressed separately.

Patch-set: 7