Update patch set 7
Patch Set 7:

> Patch Set 7: Code-Review-1
> 
> After discussion we would like this done by creating an object using the components.yml file. See the implementation of get Anchor for example. We can discuss in daily, sorry for the confusion

agreed. Not really your fault but we need to improve on the process: each simple for each user story should be approved by CPOS Mian team (or one or senior developers from team). I know we had an initial review for all you changes combined but the main feedback form that as to break it down and I implied/expected but not clearly stated (my bad) I was expecting individual reviews of each proposal going forward to discuss and agree the kind of detail that is the problem here. It is better for all involved this gets done earlier then the code review phase :-)

Patch-set: 7
Reviewer: Gerrit User 6639 <6639@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>