Update patch set 16
Patch Set 16: Code-Review-1

(2 comments)

Discussed with Toine and Daniel, we think implementation in patchset 13 is sound but the testing needs to be improved in the integration test package to account for the edge cases. Then the implementation can be fixed to match. The implementation at the db level is preferred

Patch-set: 16
Reviewer: Gerrit User 7830 <7830@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1
Work-in-progress: false