Some performance tests are skipped due to filename
CPS POM is setup so that only files ending in 'Spec' or 'Test' are
run as tests, and files ending in 'PerfTest' will only run in the
'include-performance' profile. The 32K limit tests are currently
being skipped always, as they do not end in 'Spec', 'Test', or
'PerfTest'. This commit renames them to PerfTest.

Issue-ID: CPS-1811
Signed-off-by: danielhanrahan <daniel.hanrahan@est.tech>
Change-Id: I75a4ae92edf324213f734503c23dedb2b31fd007