Remove inefficient saveListElementsBatch API
CpsDataService::saveListElementsBatch method is not needed as
saveListElements supports saving multiple list elements in a single
operation. This both simplifies implementation and greatly improves
performance when saving list elements, as the Yang parser need only
run once for the whole batch.

There is no need to deprecate the method, as it was never publicly
exposed in REST APIs, and is only used internally in NCMP.

- Change InventoryPersistence to save CM-handles in batches of 100
  using existing CpsDataService::saveListElements method.
- Remove not needed CpsDataService::saveListElementBatch.

Issue-ID: CPS-2019
Signed-off-by: danielhanrahan <daniel.hanrahan@est.tech>
Change-Id: I8b74dda2917e094d064b42f2c0e4d57029b90395