Update of top-level container node fails with SQL error
The error is caused by the fetch-descendants code using
resultSet.getLong which returns a 'long' primitive.
Thus a parent ID of 'NULL' gets translated to '0',
which can later not be found during update. To preserve
the NULL value, resultSet.getObject must be used.
A test has been added to prevent regression.

Issue-ID: CPS-1841
Signed-off-by: danielhanrahan <daniel.hanrahan@est.tech>
Change-Id: Ida6837a04954cd3c23f0f2faabd7d0712d8ee19d