Fix timing issue in CmHandleCreateSpec
The test for module sync retry on LOCKED handles has a timing issue,
due the ordering of the steps. After a LOCKED handle gets reset to
ADVISED state (but DMI is not yet set to available), the Module Sync
Watchdog may wake up and put it back into LOCKED state again.

This fix is to first make the DMI available before waiting for the
LOCKED handle to reset.

Issue-ID: CPS-2304
Signed-off-by: danielhanrahan <daniel.hanrahan@est.tech>
Change-Id: I9d33b0eeea1e0a2e163489130c928b4d03182774