Fix failing CSIT and add unit test proving the bug
It was determined that one CSIT is intermittently failing due
to a CM-handle being deleted while module sync is in progress,
which causes the whole batch operation to fail. Even CM-handles
that did sync will not go into READY state, despite the logs
saying otherwise. This commit reproduces the issue in a unit test,
and prevents the issue in the CSIT by changing test order. Also,
errors during module sync tasks are reported at ERROR level.
(The actual bug fix will be addressed in another patch.)

Issue-ID: CPS-2474
Signed-off-by: danielhanrahan <daniel.hanrahan@est.tech>
Change-Id: I7e0d617cbd48d8fd1fad036079fbd876ee21d8a8