[Cps Path Parser] Don't quote numbers in comparison
Presently, Cps Path Parser will wrap numbers in quotes when
normalizing or getting parent path, which can lead to invalid paths.
For example:
  /books[@price > 15]
will be normalized to:
  /books[@price>'15']

This causes trouble when running queries on normalized paths:
(CpsPathException: "can use only > with integer")

Note this patch will still normalize numbers to string for equality:
  /books[@price='15']
Much existing code relies on this, and it does not cause problems.

Issue-ID: CPS-2365
Signed-off-by: danielhanrahan <daniel.hanrahan@est.tech>
Change-Id: I15d326ad7db2311b64c636283a4d52623a125176