Task 3: Get delta between an anchor and JSON payload with grouping enabled
Remove @Ignore from the third Delta Report performance test (Get delta between
an anchor and JSON payload with grouping enabled) Jenkins build remains stable
or fails due to this test.

Purpose:
This change is part of an investigation to identify which Delta Report performance
test is causing the build failures.

Issue-ID: CPS-3033
Change-Id: I17ed5a52657675e9e2d1c0d75033246f09715060
Signed-off-by: Arpit Singh <AS00745003@techmahindra.com>